pyproject.toml
[tool.poetry]
name = "mcs_manager"
version = "0.1.0"
description = ""
authors = ["Austin Hacker"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
